# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wood Trading API",
        "description": "API for Wood Trading Application - A comprehensive system for managing wood inventory, sales, purchases, and financial operations.",
        "contact": {
            "name": "Wood Trading Support",
            "email": "support@woodtrading.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.woodtrading.com/api",
            "description": "Production Server"
        },
        {
            "url": "D:\\Haleem-dev\\wood-trading"
        }
    ],
    "paths": {
...
        "/api/items": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get list of items",
                "operationId": "53e44209d7fe56acf154d9c0d47e5069",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for name, code, or species",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "species",
                        "in": "query",
                        "description": "Filter by species",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grade",
                        "in": "query",
                        "description": "Filter by grade",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ItemResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Create a new item",
                "operationId": "371ad8f4244d53c482a14de9004905a4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "species",
                                    "grade",
                                    "thickness",
                                    "width",
                                    "length",
                                    "unit"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "example": "PINE-2X4-8"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Pine 2x4 8ft"
                                    },
                                    "species": {
                                        "type": "string",
                                        "example": "Pine"
                                    },
                                    "grade": {
                                        "type": "string",
                                        "example": "A"
                                    },
                                    "thickness": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 38.1
                                    },
                                    "width": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 88.9
                                    },
                                    "length": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2438.4
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "piece"
                                    },
                                    "barcode": {
                                        "type": "string",
                                        "example": "1234567890123"
                                    },
                                    "moisture_level": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12
                                    },
                                    "low_stock_threshold": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "costing_method": {
                                        "type": "string",
                                        "enum": [
                                            "FIFO",
                                            "Average"
                                        ],
                                        "example": "FIFO"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/items/{id}": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Get item by ID",
                "operationId": "cf72b0674e86263c9f13097ae1198063",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Update item",
                "operationId": "5946ad0f76a90167421f69b2ba2a086a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Pine 2x4 8ft"
                                    },
                                    "species": {
                                        "type": "string",
                                        "example": "Pine"
                                    },
                                    "grade": {
                                        "type": "string",
                                        "example": "A"
                                    },
                                    "thickness": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 38.1
                                    },
                                    "width": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 88.9
                                    },
                                    "length": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2438.4
                                    },
                                    "unit": {
                                        "type": "string",
                                        "example": "piece"
                                    },
                                    "barcode": {
                                        "type": "string",
                                        "example": "1234567890123"
                                    },
                                    "moisture_level": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12
                                    },
                                    "low_stock_threshold": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "costing_method": {
                                        "type": "string",
                                        "enum": [
                                            "FIFO",
                                            "Average"
                                        ],
                                        "example": "FIFO"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ItemResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Inventory"
                ],
                "summary": "Delete item",
                "operationId": "264a2ea4d554347870bc1216f33f30a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Item deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ItemResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "example": "WOOD001"
                    },
                    "name": {
                        "type": "string",
                        "example": "Oak Hardwood Plank"
                    },
                    "species": {
                        "type": "string",
                        "example": "Oak"
                    },
                    "grade": {
                        "type": "string",
                        "example": "A"
                    },
                    "thickness": {
                        "type": "number",
                        "format": "float",
                        "example": 2.5
                    },
                    "width": {
                        "type": "number",
                        "format": "float",
                        "example": 6
                    },
                    "length": {
                        "type": "number",
                        "format": "float",
                        "example": 96
                    },
                    "unit": {
                        "type": "string",
                        "example": "pieces"
                    },
                    "barcode": {
                        "type": "string",
                        "example": "1234567890123"
                    },
                    "moisture_level": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "low_stock_threshold": {
                        "type": "integer",
                        "example": 100
                    },
                    "costing_method": {
                        "type": "string",
                        "example": "FIFO"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "total_stock": {
                        "type": "integer",
                        "example": 500
                    },
                    "available_stock": {
                        "type": "integer",
                        "example": 450
                    },
                    "reserved_stock": {
                        "type": "integer",
                        "example": 50
                    },
                    "is_low_stock": {
                        "type": "boolean",
                        "example": false
                    },
                    "stock_lots": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/StockLotResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "StockLotResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lot_no": {
                        "type": "string",
                        "example": "LOT2024001"
                    },
                    "qty_on_hand": {
                        "type": "integer",
                        "example": 100
                    },
                    "qty_reserved": {
                        "type": "integer",
                        "example": 10
                    },
                    "qty_available": {
                        "type": "integer",
                        "example": 90
                    },
                    "cost": {
                        "type": "number",
                        "format": "float",
                        "example": 25.5
                    },
                    "received_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "warehouse": {
                        "$ref": "#/components/schemas/WarehouseResource"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "WarehouseResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Warehouse"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Industrial Blvd, City, State"
                    },
                    "contact_person": {
                        "type": "string",
                        "example": "Jane Doe"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1-555-0456"
                    },
                    "email": {
                        "type": "string",
                        "example": "warehouse@company.com"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}